chore: remove old tables
