+++ content/blocks-components.tsx
@@ -65,8 +65,7 @@ export const blocksComponents: { [blocksId: string]: React.ElementType } = {
   "ai-02": components.Ai02,
   "ai-03": components.Ai03,
   "ai-04": components.Ai04,
+
   "table-01": components.Table01,
   "table-02": components.Table02,
-  "table-03": components.Table03,
-  "table-04": components.Table04,
 };
