+++ content/blocks-metadata.ts
@@ -411,26 +411,12 @@ export const blocksMetadata: BlocksMetadata[] = [
   {
     id: "table-01",
     category: categoryIds.Tables,
-    name: "Simple Data Table",
-    iframeHeight: "400px",
-    type: "file",
-  },
-  {
-    id: "table-02",
-    category: categoryIds.Tables,
-    name: "Orders Data Table",
-    iframeHeight: "600px",
-    type: "file",
-  },
-  {
-    id: "table-03",
-    category: categoryIds.Tables,
     name: "Accordion Table",
     iframeHeight: "800px",
     type: "file",
   },
   {
-    id: "table-04",
+    id: "table-02",
     category: categoryIds.Tables,
     name: "Task Management Table",
     iframeHeight: "600px",
