+++ content/components/tables/index.ts
@@ -1,5 +1,3 @@
 // Export your components here
 export { default as Table01 } from "./table-01";
 export { default as Table02 } from "./table-02";
-export { default as Table03 } from "./table-03";
-export { default as Table04 } from "./table-04";
